# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchUploadItemOut": {
        "description": "批量上传的逐文件结果：部分成功语义——坏文件只在它自己那行有 error。",
        "properties": {
          "document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "document",
          "error"
        ],
        "title": "BatchUploadItemOut",
        "type": "object"
      },
      "Body_upload_document_api_v1_kb__kb_id__documents_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_kb__kb_id__documents_post",
        "type": "object"
      },
      "Body_upload_documents_batch_api_v1_kb__kb_id__documents_batch_post": {
        "properties": {
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_upload_documents_batch_api_v1_kb__kb_id__documents_batch_post",
        "type": "object"
      },
...
      "ChunkPreviewOut": {
        "properties": {
          "chunk_index": {
            "title": "Chunk Index",
            "type": "integer"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "has_embedding": {
            "title": "Has Embedding",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "id",
          "chunk_index",
          "content",
          "has_embedding",
          "meta"
        ],
        "title": "ChunkPreviewOut",
        "type": "object"
      },
...
      "DocOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "has_embedding": {
            "title": "Has Embedding",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kb_id": {
            "title": "Kb Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "status": {
            "enum": [
              "pending",
              "parsing",
              "ready",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kb_id",
          "name",
          "status",
          "error",
          "size_bytes",
          "created_at",
          "has_embedding"
        ],
        "title": "DocOut",
        "type": "object"
      },
      "DocPatchIn": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "pending",
                  "parsing",
                  "ready",
                  "failed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "DocPatchIn",
        "type": "object"
      },
      "ErrorOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Umax RAG",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/{doc_id}": {
      "delete": {
        "description": "删文档：切块级联清、落盘原文件一并删（别把客户磁盘当垃圾场）。\n\n不可逆（要恢复得重新上传重建索引），故前端有二次确认；前端/接口的\n消息引用是历史快照，不随文档删除而变（当时的回答就该保持当时的出处）。",
        "operationId": "delete_document_api_v1_documents__doc_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": -2147483648,
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "需要登录"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "需要管理员权限"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "文档不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "首次登录必须修改初始口令"
          }
        },
        "summary": "Delete Document"
      },
      "get": {
        "operationId": "get_document_api_v1_documents__doc_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": -2147483648,
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "需要登录"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "文档不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "首次登录必须修改初始口令"
          }
        },
        "summary": "Get Document"
      },
      "patch": {
        "operationId": "patch_document_api_v1_documents__doc_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": -2147483648,
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocPatchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "请求体解析失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "需要登录"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "需要管理员权限"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "文档不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "首次登录必须修改初始口令"
          }
        },
        "summary": "Patch Document"
      }
    },
    "/api/v1/documents/{doc_id}/chunks": {
      "get": {
        "operationId": "preview_chunks_api_v1_documents__doc_id__chunks_get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": -2147483648,
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChunkPreviewOut"
                  },
                  "title": "Response Preview Chunks Api V1 Documents  Doc Id  Chunks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "需要登录"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "文档不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "首次登录必须修改初始口令"
          }
        },
        "summary": "Preview Chunks"
      }
    },
    "/api/v1/documents/{doc_id}/reprocess": {
      "post": {
        "operationId": "reprocess_api_v1_documents__doc_id__reprocess_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": -2147483648,
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "需要登录"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "需要管理员权限"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "文档不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "首次登录必须修改初始口令"
          }
        },
        "summary": "Reprocess"
      }
    },
...
    "/api/v1/kb/{kb_id}/documents": {
      "get": {
        "operationId": "list_documents_api_v1_kb__kb_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": -2147483648,
              "title": "Kb Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocOut"
                  },
                  "title": "Response List Documents Api V1 Kb  Kb Id  Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "需要登录"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "知识库不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "首次登录必须修改初始口令"
          }
        },
        "summary": "List Documents"
      },
      "post": {
        "operationId": "upload_document_api_v1_kb__kb_id__documents_post",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": -2147483648,
              "title": "Kb Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_kb__kb_id__documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "请求体解析失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "需要登录"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "需要管理员权限"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "知识库不存在"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "不支持的文件类型"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "首次登录必须修改初始口令"
          }
        },
        "summary": "Upload Document"
      }
    },
    "/api/v1/kb/{kb_id}/documents/batch": {
      "post": {
        "description": "批量上传（§A）：逐文件给出结果，**部分成功**——坏文件只在自己那行报错。\n\n与单文件端点的差异（有意）：不支持的扩展名不再整批 415，而是该行 error 字段。",
        "operationId": "upload_documents_batch_api_v1_kb__kb_id__documents_batch_post",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": -2147483648,
              "title": "Kb Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_batch_api_v1_kb__kb_id__documents_batch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BatchUploadItemOut"
                  },
                  "title": "Response Upload Documents Batch Api V1 Kb  Kb Id  Documents Batch Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "请求体解析失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "需要登录"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "需要管理员权限"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "知识库不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "首次登录必须修改初始口令"
          }
        },
        "summary": "Upload Documents Batch"
      }
    },
...
  }
}